Move dataclasses from controller/__init__ to data_model
Route, BackgroundRSSI, RFRegionInfo, NVMReadResult, NVMOpenExtResult,
and NVMProgress now live in data_model.py alongside the existing
ZWaveChipType and TypedDict definitions. Controller.__init__ imports
them for use in method signatures. Tests import from data_model
directly rather than relying on re-exports.
