+++ zwave_js_server/model/controller/__init__.py
@@ -2,7 +2,6 @@
 
 from __future__ import annotations
 
-from dataclasses import dataclass
 import logging
 from typing import TYPE_CHECKING, Any, Literal, cast
 
@@ -29,7 +28,17 @@ from ...util.helpers import convert_base64_to_bytes, convert_bytes_to_base64
 from ..association import AssociationAddress, AssociationGroup
 from ..node import Node
 from ..node.firmware import NodeFirmwareUpdateResult
-from .data_model import ControllerDataType, ZWaveApiVersionDataType, ZWaveChipType
+from .data_model import (
+    BackgroundRSSI,
+    ControllerDataType,
+    NVMOpenExtResult,
+    NVMProgress,
+    NVMReadResult,
+    RFRegionInfo,
+    Route,
+    ZWaveApiVersionDataType,
+    ZWaveChipType,
+)
 from .event_model import CONTROLLER_EVENT_MODEL_MAP
 from .inclusion_and_provisioning import (
     InclusionGrant,
@@ -65,73 +74,6 @@ DEFAULT_CONTROLLER_STATISTICS = (  # pylint: disable=invalid-name
         timeoutCallback=0,
     )
 )
-
-
-@dataclass(frozen=True)
-class Route:
-    """A Z-Wave routing instruction specifying repeaters and speed."""
-
-    repeaters: list[Node]
-    route_speed: int
-
-    def to_dict(self) -> dict[str, Any]:
-        """Return wire-format dict."""
-        return {
-            "repeaters": [n.node_id for n in self.repeaters],
-            "routeSpeed": self.route_speed,
-        }
-
-    @classmethod
-    def from_dict(cls, data: dict, nodes: dict[int, Node]) -> Route:
-        """Initialize from wire-format dict."""
-        return cls(
-            repeaters=[nodes[nid] for nid in data["repeaters"]],
-            route_speed=data["routeSpeed"],
-        )
-
-
-@dataclass(frozen=True)
-class BackgroundRSSI:
-    """Background RSSI noise levels for all channels."""
-
-    channel_0: int
-    channel_1: int
-    channel_2: int | None = None
-    channel_3: int | None = None
-
-
-@dataclass(frozen=True)
-class RFRegionInfo:
-    """Information about an RF region."""
-
-    region: int
-    supports_zwave: bool
-    supports_long_range: bool
-    includes_region: int | None = None
-
-
-@dataclass(frozen=True)
-class NVMReadResult:
-    """Result of an NVM buffer read operation."""
-
-    buffer: bytes
-    end_of_file: bool
-
-
-@dataclass(frozen=True)
-class NVMOpenExtResult:
-    """Result of an extended NVM open operation."""
-
-    size: int
-    supported_operations: list
-
-
-@dataclass
-class NVMProgress:
-    """Class to represent an NVM backup/restore progress event."""
-
-    bytes_read_or_written: int
-    total_bytes: int
 
 
 class Controller(EventBase):
