+++ zwave_js_server/model/controller/data_model.py
@@ -3,9 +3,12 @@
 from __future__ import annotations
 
 from dataclasses import dataclass
-from typing import Literal, TypedDict
+from typing import TYPE_CHECKING, Any, Literal, TypedDict
 
 from .statistics import ControllerStatisticsDataType
+
+if TYPE_CHECKING:
+    from ..node import Node
 
 
 class ZWaveApiVersionDataType(TypedDict):
@@ -40,6 +43,73 @@ class ZWaveChipType:
         if isinstance(data, str):
             return cls(name=data)
         return cls(type=data["type"], version=data["version"])
+
+
+@dataclass(frozen=True)
+class Route:
+    """A Z-Wave routing instruction specifying repeaters and speed."""
+
+    repeaters: list[Node]
+    route_speed: int
+
+    def to_dict(self) -> dict[str, Any]:
+        """Return wire-format dict."""
+        return {
+            "repeaters": [n.node_id for n in self.repeaters],
+            "routeSpeed": self.route_speed,
+        }
+
+    @classmethod
+    def from_dict(cls, data: dict, nodes: dict[int, Node]) -> Route:
+        """Initialize from wire-format dict."""
+        return cls(
+            repeaters=[nodes[nid] for nid in data["repeaters"]],
+            route_speed=data["routeSpeed"],
+        )
+
+
+@dataclass(frozen=True)
+class BackgroundRSSI:
+    """Background RSSI noise levels for all channels."""
+
+    channel_0: int
+    channel_1: int
+    channel_2: int | None = None
+    channel_3: int | None = None
+
+
+@dataclass(frozen=True)
+class RFRegionInfo:
+    """Information about an RF region."""
+
+    region: int
+    supports_zwave: bool
+    supports_long_range: bool
+    includes_region: int | None = None
+
+
+@dataclass(frozen=True)
+class NVMReadResult:
+    """Result of an NVM buffer read operation."""
+
+    buffer: bytes
+    end_of_file: bool
+
+
+@dataclass(frozen=True)
+class NVMOpenExtResult:
+    """Result of an extended NVM open operation."""
+
+    size: int
+    supported_operations: list
+
+
+@dataclass
+class NVMProgress:
+    """Class to represent an NVM backup/restore progress event."""
+
+    bytes_read_or_written: int
+    total_bytes: int
 
 
 class ControllerDataType(TypedDict, total=False):
